마이페이지 리퀘스트

+++ app/src/main/java/com/weddingsingers/wsapp/main/MainActivity.java
@@ -41,7 +41,7 @@ import butterknife.ButterKnife;
 public class MainActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener {
 
-    public static int isLogin = 0;
+    public static long isLogin = 0;
 
     public static final int MESSAGE_BACK_KEY_TIMEOUT = 1;
     public static final int TIMEOUT_TIME = 2000;
@@ -112,7 +112,7 @@ public class MainActivity extends AppCompatActivity
         int fragmentName = intent.getIntExtra(FRAG_NAME, ERROR_CODE);
         changeFragmentFromAnotherActivity(fragmentName);
 
-        isLogin = intent.getIntExtra(EXTRA_USER_ID, 0);
+        isLogin = intent.getLongExtra(EXTRA_USER_ID, 0);
         if(isLogin == 0) {
             checkLogin(false);
         } else {
